backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
firebase-admin==6.5.0
insightface==0.7.3
opencv-python==4.10.0.84
numpy==1.26.4
onnxruntime==1.19.1
python-dotenv==1.0.1
onnxruntime-gpu==1.19.1; sys_platform != 'darwin'  # GPU support (optional, comment out if not using GPU)
resemblyzer==0.1.1
scipy==1.11.4
soundfile==0.12.1
sinch-python-sdk==2.0.0
requests==2.31.0

